+++ release_notes.md
@@ -6,6 +6,9 @@
 
 ## Code
 
+- Fixed vscode project
+- Code cleanup on embedded console
+- Test case for EventHub
 - BrokerLib: Thinkers linked list now is double linked for fast removal
 - Common: Fixed Socket move assignement
 - BrokerLib: EventHub now has a pool and internal list, avoid excessive mem alloc.